requirements.txt
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
gitdb==4.0.10
GitPython==3.1.31
ipykernel==6.21.2
ipython==8.10.0
jedi==0.18.2
jupyter_client==8.0.3
jupyter_core==5.2.0
lxml==4.9.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
wcwidth==0.2.6
